Organize country helper methods.
