+++ backend/app/views/spree/admin/zones/_country_members.html.erb
@@ -9,7 +9,7 @@
     <div class="card-body">
       <%= zone_form.field_container :country_ids, class: ['form-group'] do %>
         <%= zone_form.label :country_ids, Spree.t(:countries) %>
-        <%= zone_form.collection_select :country_ids, all_country_options(@countries), :last, :first, {}, { multiple: true, class: "select2" } %>
+        <%= zone_form.collection_select :country_ids, all_country_options_localized, :last, :first, {}, { multiple: true, class: "select2" } %>
       <% end %>
     </div>
   </div>
