+++ core/app/helpers/spree/base_helper.rb
@@ -1,18 +1,5 @@
 module Spree
   module BaseHelper
-    include Spree::LocaleHelper
-    include Spree::CountryHelper
-
-    def available_countries
-      countries = current_store.countries_available_for_checkout
-
-      countries.collect do |country|
-        country.name = localized_country_name(country.iso)
-
-        country
-      end.sort_by { |c| c.name.parameterize }
-    end
-
     def display_price(product_or_variant)
       product_or_variant.
         price_in(current_currency).
