+++ core/spec/helpers/base_helper_spec.rb
@@ -10,67 +10,6 @@ describe Spree::BaseHelper, type: :helper do
     allow(controller).to receive(:controller_name).and_return('test')
   end
 
-  context 'available_countries' do
-    let(:country) { create(:country, name: 'United States', iso: 'US') }
-    let(:country_1) { create(:country, name: 'Germany', iso: 'DE') }
-    let(:country_2) { create(:country, name: 'England', iso: 'GB') }
-    let(:country_3) { create(:country, name: 'France', iso: 'FR') }
-
-    context 'with checkout zone assigned to the store' do
-      before do
-        Spree::Config[:checkout_zone] = nil
-        @zone = create(:zone, name: 'No Limits', kind: 'country')
-        @zone.members.create(zoneable: country)
-        current_store.update(checkout_zone_id: @zone.id)
-      end
-
-      it 'return only the countries defined by the checkout_zone_id' do
-        expect(available_countries).to eq([country])
-        expect(current_store.checkout_zone_id).to eq @zone.id
-      end
-    end
-
-    context 'with no checkout zone defined' do
-      before do
-        Spree::Config[:checkout_zone] = nil
-        current_store.update(checkout_zone_id: nil)
-      end
-
-      it 'return complete list of countries' do
-        expect(available_countries.count).to eq(Spree::Country.count)
-      end
-    end
-
-    context 'with a checkout zone defined' do
-      context 'checkout zone is of type country' do
-        before do
-          @country_zone = create(:zone, name: 'CountryZone', kind: 'country')
-          @country_zone.members.create(zoneable: country)
-          Spree::Config[:checkout_zone] = @country_zone.name
-        end
-
-        it 'return only the countries defined by the checkout zone' do
-          expect(available_countries).to eq([country])
-        end
-      end
-
-      context 'checkout zone is of type state' do
-        let(:state) { create(:state, country: country) }
-
-        before do
-          state_zone = create(:zone, name: 'StateZone')
-          state_zone.members.create(zoneable: state)
-
-          Spree::Config[:checkout_zone] = state_zone.name
-        end
-
-        it 'returns list of countries associated with states' do
-          expect(available_countries).to contain_exactly state.country
-        end
-      end
-    end
-  end
-
   # Regression test for #1436
   context 'defining custom image helpers' do
     let(:product) { build(:product) }
